refactor(models): simplify transformer ensemble architecture
Replace custom LinearEnsemble-based block structure with standard
TransformerEnsembleBlock instances. The forward pass now applies
blocks directly without the custom residual connection pattern and
transpose operations.

Data preprocessing pipeline is adjusted to match:
- Remove KBinsDiscretizer and custom NaN filling
- Enable QuantileTransformer for normalization
- x_num embedding dimension changed from n_bins to raw feature count
- Numerical features cast to float32 instead of long

BREAKING CHANGE: Model architecture and data preprocessing have changed
significantly. Previous checkpoints and configurations using discretized
features are incompatible. The model now expects continuous numerical
features directly.

+++ data/custom_dataset/custom_dataset.py
@@ -14,7 +14,7 @@ class CustomDataset(Dataset):
                  target: np.ndarray,
                  label: np.ndarray
                  ) -> None:
-        self.x_num = torch.as_tensor(x_num, dtype=torch.long)
+        self.x_num = torch.as_tensor(x_num, dtype=torch.float32)
         self.x_cat = torch.as_tensor(x_cat, dtype=torch.long)
         self.target = torch.as_tensor(target, dtype=torch.float32)
         # print(label)
